requirements.txt
aiogram
fastapi
uvicorn
python-dotenv
psycopg2-binary
requests
